\name{erer-package}
\alias{erer}
\alias{erer-package}
\docType{package}
\title{Empirical Research in Economics with R }
\description{
This package contains functions and datasets for the book of 'Empirical Research in Economics: Growing up with R' by Dr. Changyou Sun. These functions can calculate marginal effects for a binary probit or logit model; calculate marginal effects for an ordered probit or ordered logit model; estimate static and dynamic Almost Ideal Demand System (AIDS) models; and conduct event analysis.
}
\details{
\tabular{ll}{
Package:  \tab erer \cr
Type:     \tab Package \cr
Version:  \tab 2.1 \cr
Date:     \tab 2010-11-15 (first built); 2014-6-3 (last) \cr
Depends:  \tab R (>= 3.0.0), lmtest \cr
Imports:  \tab systemfit, tseries, urca, ggplot2 \cr
Suggests: \tab MASS \cr
License:  \tab GPL (>= 2) \cr
LazyLoad: \tab yes \cr
}
}
\author{ Changyou Sun <csun@cfr.msstate.edu>}
\keyword{package} 