\name{print.evReturn}
\alias{print.evReturn}
\title{Printing Abnormal Return Results from Event Analsyis}
\description{Show abnormal return resutls from \code{evReturn}.}
\usage{\method{print}{evReturn}(x, \dots)}
\arguments{
  \item{x}{an object of class \code{evReturn}.}
  \item{\dots}{additional arguments to be passed.}
}
\details{This print method for object of class \code{evReturn}.}
\value{Summary results of the coefficents or outputs.}
\author{Changyou Sun (\email{csun@cfr.msstate.edu})}
\seealso{\code{\link{evReturn}}; \code{\link{plot.evReturn}}.}

\examples{
# see the examples for 'evReturn'.
}
\keyword{methods}