\name{print.evRisk}
\alias{print.evRisk}
\title{Printing Risk Results from Event Analsyis}
\description{Show regression resutls from \code{evRisk}.}
\usage{\method{print}{evRisk}(x, \dots)}
\arguments{
  \item{x}{an object of class \code{evRisk}.}
  \item{\dots}{additional arguments to be passed.}
}
\details{This print method for object of class \code{evRisk}.}
\value{Summary results of the coefficents or outputs.}
\author{Changyou Sun (\email{csun@cfr.msstate.edu})}
\seealso{\code{\link{evRisk}}.}

\examples{
# see the examples for 'evRisk'.
}
\keyword{methods}