\name{erer-package}
\alias{erer}
\alias{erer-package}
\docType{package}
\title{Empirical Research in Economics with R }
\description{
Functions, datasets, and sample codes related to the book of 'Empirical Research in Economics: Growing up with R' by Dr. Changyou Sun are included. The book is still under preparation as of August 2014. Marginal effects for binary or ordered choice models can be calculated. Static and dynamic Almost Ideal Demand System (AIDS) models can be estimated. A typical event analysis in finance can be conducted with several functions included.

There are over 80 sample programs for the book. They are included in the package folder on your local drive, e.g., c:/.../R-3.1.1/library/erer/doc/ererBookRcode. Alternatively, download the tar.gz version from the CRAN site to view them.
}
\details{
\tabular{ll}{
Package:  \tab erer \cr
Type:     \tab Package \cr
Version:  \tab 2.2 \cr
Date:     \tab 2010-11-15 (first built); 2014-8-10 (last) \cr
Depends:  \tab R (>= 3.0.0), lmtest \cr
Imports:  \tab systemfit, tseries, urca, ggplot2 \cr
Suggests: \tab MASS \cr
License:  \tab GPL (>= 2) \cr
LazyLoad: \tab yes \cr
}
}
\author{ Changyou Sun <csun@cfr.msstate.edu>}
\keyword{package} 