\name{erer-package}
\alias{erer}
\alias{erer-package}
\docType{package}
\title{Empirical Research in Economics with R }
\description{
This package contains functions and datasets for the book of 'Empirical Research in Economics: Growing up with R' by Dr. Changyou Sun. These functions can calculate marginal effects for a binary probit or logit model, estimate static and dynamic Almost Ideal Demand System (AIDS) models, and conduct event analysis.
}
\details{
\tabular{ll}{
Package:  \tab erer \cr
Type:     \tab Package \cr
Version:  \tab 1.0 \cr
Date:     \tab 2010-11-15 (first built) \cr
Depends:  \tab R (>= 2.10.0), systemfit, lmtest, tseries, ggplot2 \cr
License:  \tab GPL (>=2) \cr
LazyLoad: \tab yes \cr
}
}
\author{ Changyou Sun <csun@cfr.msstate.edu>}
\keyword{package} 