\name{erer-package}
\alias{erer}
\alias{erer-package}
\docType{package}
\title{Empirical Research in Economics with R }
\description{
Functions, datasets, and sample codes related to the book of 'Empirical Research in Economics: Growing up with R' by Dr. Changyou Sun are included (ISBN 9780996585408). Marginal effects for binary or ordered choice models can be calculated. Static and dynamic Almost Ideal Demand System (AIDS) models can be estimated. A typical event analysis in finance can be conducted with several functions included.

Close to 100 sample programs and data sets are used in the book. They are included in the package folder on your local drive, e.g., c:/.../R-3.2.2/library/erer/doc/. Alternatively, download the tar.gz version of this package from the CRAN site to view them.
}
\details{
\tabular{ll}{
Package:  \tab erer \cr
Type:     \tab Package \cr
Version:  \tab 3.1 \cr
Date:     \tab 2022-04-18 \cr
Depends:  \tab R (>= 3.0.0), lmtest \cr
Imports:  \tab systemfit, tseries, urca \cr
Suggests: \tab MASS \cr
License:  \tab GPL (>= 2) \cr
LazyLoad: \tab yes \cr
}
}
\author{Changyou Sun (\email{cs258@msstate.edu})}
\keyword{package} 